Use likelihood attributes in NCE + Fix System Version + SVC improvements
The SVCs improvements are as follows:
* Make SVC logs more concise for:
* * `SleepThread`
* * `ClearEvent`
* * `CloseHandle`
* * `ResetSignal`
* * `WaitSynchronization` (Special case for single handle)
* * `ArbitrateLock`
* * `ArbitrateUnlock`
* * `WaitProcessWideKeyAtomic`
* * `SignalProcessWideKey`
* Fix unintentional fallthrough into `yieldWithoutCoreMigration` from `yieldWithCoreMigration` in `SleepThread`
* Return `result::InvalidState` when an unsignalled handle is reset in `ResetSignal`
* Return `Result{}` (Success) in `CancelSynchronization`
* Do not return `result::InvalidCurrentMemory` in `ArbitrateLock` as it's not a failure condition
* Make `count` in `WaitProcessWideKeyAtomic` a `i32` from a `u32`, zero and all negative values result in waking all waiters

+++ app/src/main/cpp/skyline/kernel/types/KProcess.h
@@ -221,7 +221,7 @@ namespace skyline {
             /**
              * @brief Signals the conditional variable at the specified address
              */
-            void ConditionalVariableSignal(u32 *key, u64 amount);
+            void ConditionalVariableSignal(u32 *key, i32 amount);
 
             /**
              * @brief Waits on the supplied address with the specified arbitration function
